Make release dependent module loading static
Currently the ClasspathJrt uses a static loadModules method while the
ClasspathJrtWithReleaseOption uses an instance method (that effectively
modifies static state). This makes it harder to reuse and is prone to
changes in the handling of instance fields.

This now extracts the part into the base class ClasspathJrt and passing
the CtSym and release code as a parameter instead and streamlining the
logic if modules should be loaded or not to allow reuse and easier
debugging.
